Added an option for source term diffusion to the cloud
setting in the solution subdict in the cloud dictionary:

    sourceDistribution true;

    diffusion
    {
        useImplicitLaplacian true;
        //proposed to be 6*dp
        // can be set separately for the different source
        // terms (momentum, radia, etc.) default like this
        // means same bandwidth for all
        diffusionBandWidth  180e-5;
    }

additionally the following ddtScheme needs to be set:

    "ddt(Smooth)"     independentEuler;

and the custom library needs to be included to make it available.

The diffusion is based on this work:

Zhang J, Li T, Ström H, et al.
Grid-independent Eulerian-Lagrangian approaches for simulations of solid fuel particle combustion.
Chemical Engineering Journal, 2020, 387: 123964.